Fixed 0x00EE Fixed 0x2nnn Made stack 16 bits wide.

+++ src/Memory.cpp
@@ -71,10 +71,10 @@ void Memory::reset() {
     writeBytes(this->_fontset, 80, Memory::fontsetStartaddress);
 }
 
-uint8_t Memory::popStack(uint16_t sp) {
+uint16_t Memory::popStack(uint16_t sp) {
     return this->_stack[sp];
 }
 
-void Memory::pushStack(const uint16_t sp, const uint8_t value) {
+void Memory::pushStack(const uint16_t sp, const uint16_t value) {
     this->_stack[sp] = value;
 }
